Rename module to api
