+++ foxglove/client/__init__.py
@@ -1,3 +1,3 @@
-from .client import Client, FoxgloveException, OutputFormat
+from .api import Client, FoxgloveException, OutputFormat
 
 __all__ = ["Client", "FoxgloveException", "OutputFormat"]
